+++ m-common/src/main/java/com/antmen/antwork/common/api/controller/CommonReviewController.java
@@ -1,0 +1,42 @@
+package com.antmen.antwork.common.api.controller;
+
+import com.antmen.antwork.common.api.request.reservation.ReviewRequestDto;
+import com.antmen.antwork.common.api.response.reservation.ReviewResponseDto;
+import com.antmen.antwork.common.service.serviceReservation.ReviewService;
+import jakarta.validation.Valid;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/api/v1/common/reviews")
+public class CommonReviewController {
+    private final ReviewService reviewService;
+
+    // 리뷰 단건 조회
+    @GetMapping("/{reviewId}")
+    public ResponseEntity<ReviewResponseDto> getReview(
+            @PathVariable
+            Long reviewId
+    ) {
+        return ResponseEntity.ok(reviewService.getReview(reviewId));
+    }
+
+    // 리뷰 전체 조회
+    @GetMapping
+    public ResponseEntity<List<ReviewResponseDto>> getAllReviews() {
+        return ResponseEntity.ok(reviewService.getAllReviews());
+    }
+
+    // 특정 id의 리뷰 목록 조회
+    @GetMapping("/users/{userId}")
+    public ResponseEntity<List<ReviewResponseDto>> getReviewsByUserId(
+            @PathVariable
+            Long userId
+    ) {
+       return ResponseEntity.ok(reviewService.getReviewsByUserId(userId));
+    }
+}
