+++ m-customer/src/main/java/com/antmen/antwork/customer/controller/CustomerReviewController.java
@@ -1,4 +1,4 @@
-package com.antmen.antwork.common.api.controller;
+package com.antmen.antwork.customer.controller;
 
 import com.antmen.antwork.common.api.request.reservation.ReviewRequestDto;
 import com.antmen.antwork.common.api.response.reservation.ReviewResponseDto;
@@ -11,9 +11,9 @@ import org.springframework.web.bind.annotation.*;
 import java.util.List;
 
 @RestController
-@RequestMapping("/review")
 @RequiredArgsConstructor
-public class ReviewController {
+@RequestMapping("/api/v1/customer/reviews")
+public class CustomerReviewController {
     private final ReviewService reviewService;
 
     // 리뷰 등록
@@ -27,42 +27,11 @@ public class ReviewController {
         return ResponseEntity.ok(reviewService.createReview(loginId, dto));
     }
 
-    // 리뷰 단건 조회
-    @GetMapping("/{reviewId}")
-    public ResponseEntity<ReviewResponseDto> getReview(
-            @PathVariable
-            Long reviewId
-    ) {
-        return ResponseEntity.ok(reviewService.getReview(reviewId));
-    }
-
-    // 리뷰 전체 조회
-    @GetMapping
-    public ResponseEntity<List<ReviewResponseDto>> getAllReviews() {
-        return ResponseEntity.ok(reviewService.getAllReviews());
-    }
-
-    // 내가 받은 리뷰 목록 조회
-    @GetMapping("/received")
-    public ResponseEntity<List<ReviewResponseDto>> getMyReceivedReviews(){
-        Long loginId = 3L;
-        return ResponseEntity.ok(reviewService.getMyReceivedReviews(loginId));
-    }
-
     // 내가 쓴 리뷰 목록 조회
-    @GetMapping("/written")
+    @GetMapping("/my/written")
     public ResponseEntity<List<ReviewResponseDto>> getMyWrittenReviews() {
         Long loginId = 3L;
         return ResponseEntity.ok(reviewService.getMyWrittenReviews(loginId));
-    }
-
-    // 특정 id의 리뷰 목록 조회
-    @GetMapping("/users/{userId}")
-    public ResponseEntity<List<ReviewResponseDto>> getReviewsByUserId(
-            @PathVariable
-            Long userId
-    ) {
-       return ResponseEntity.ok(reviewService.getReviewsByUserId(userId));
     }
 
     // 리뷰 수정
